requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
delta-spark==4.0.0
geopandas==1.1.1
idna==3.11
importlib_metadata==8.7.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
py4j==0.10.9.9
pyarrow==22.0.0
pyogrio==0.11.1
pyproj==3.7.2
pyspark==4.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
scipy==1.16.2
shapely==2.1.2
six==1.17.0
tabulate==0.9.0
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0